move BasicConfig
