+++ hedera-node/hedera-config/src/testFixtures/java/com/hedera/node/config/testfixtures/HederaTestConfigBuilder.java
@@ -77,8 +77,8 @@ import com.swirlds.platform.health.OSHealthCheckConfig;
 import com.swirlds.virtualmap.config.VirtualMapConfig;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import org.hiero.base.crypto.config.CryptoConfig;
+import org.hiero.consensus.BasicConfig;
 import org.hiero.consensus.concurrent.config.BasicCommonConfig;
-import org.hiero.consensus.config.BasicConfig;
 import org.hiero.consensus.config.EventConfig;
 import org.hiero.consensus.config.PathsConfig;
 import org.hiero.consensus.config.PlatformStatusConfig;
