+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/config/UtilityConfigurationExtension.java
@@ -16,7 +16,6 @@ public class UtilityConfigurationExtension implements ConfigurationExtension {
     @NonNull
     public Set<Class<? extends Record>> getConfigDataTypes() {
         return Set.of(
-                BasicConfig.class,
                 EventConfig.class,
                 FallenBehindConfig.class,
                 PathsConfig.class,
