+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/config/PlatformConfigurationExtension.java
@@ -13,6 +13,7 @@ import com.swirlds.platform.monitor.StatusMonitorWiringConfig;
 import com.swirlds.platform.uptime.UptimeConfig;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Set;
+import org.hiero.consensus.BasicConfig;
 import org.hiero.consensus.model.node.NodeId;
 
 /**
@@ -30,6 +31,7 @@ public class PlatformConfigurationExtension implements ConfigurationExtension {
 
         // Load Configuration Definitions
         return Set.of(
+                BasicConfig.class,
                 ModulesConfig.class,
                 OSHealthCheckConfig.class,
                 PlatformMetricsConfig.class,
