+++ platform-sdk/swirlds-platform-core/src/main/java/org/hiero/consensus/ConsensusLayerStaticSetup.java
@@ -25,7 +25,6 @@ import java.nio.file.Path;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.hiero.consensus.config.BasicConfig;
 import org.hiero.consensus.config.PathsConfig;
 
 /**
